Use recommands python project organisation
Tools like find_packages() expected code source under a sub directory
generated entry point from released package were wrong.

let a patator.py in the root directory, this file won't be
in the released pacage. Its intent is for backward
compatibility.
