+++ setup.py
@@ -55,40 +55,40 @@ setup(
     url="https://github.com/lanjelot/patator",
     entry_points={
         'console_scripts': [
-            'ftp_login=patator:main',
-            'ssh_login=patator:main',
-            'telnet_login=patator:main',
-            'smtp_login=patator:main',
-            'smtp_vrfy=patator:main',
-            'smtp_rcpt=patator:main',
-            'finger_lookup=patator:main',
-            'http_fuzz=patator:main',
-            'ajp_fuzz=patator:main',
-            'pop_login=patator:main',
-            'pop_passd=patator:main',
-            'imap_login=patator:main',
-            'ldap_login=patator:main',
-            'smb_login=patator:main',
-            'smb_lookupsid=patator:main',
-            'rlogin_login=patator:main',
-            'vmauthd_login=patator:main',
-            'mssql_login=patator:main',
-            'oracle_login=patator:main',
-            'mysql_login=patator:main',
-            'mysql_query=patator:main',
-            'rdp_login=patator:main',
-            'pgsql_login=patator:main',
-            'vnc_login=patator:main',
-            'dns_forward=patator:main',
-            'dns_reverse=patator:main',
-            'snmp_login=patator:main',
-            'ike_enum=patator:main',
-            'unzip_pass=patator:main',
-            'keystore_pass=patator:main',
-            'sqlcipher_pass=patator:main',
-            'umbraco_crack=patator:main',
-            'tcp_fuzz=patator:main',
-            'dummy_test=patator:main',
+            'ftp_login=patator.patator:main',
+            'ssh_login=patator.patator:main',
+            'telnet_login=patator.patator:main',
+            'smtp_login=patator.patator:main',
+            'smtp_vrfy=patator.patator:main',
+            'smtp_rcpt=patator.patator:main',
+            'finger_lookup=patator.patator:main',
+            'http_fuzz=patator.patator:main',
+            'ajp_fuzz=patator.patator:main',
+            'pop_login=patator.patator:main',
+            'pop_passd=patator.patator:main',
+            'imap_login=patator.patator:main',
+            'ldap_login=patator.patator:main',
+            'smb_login=patator.patator:main',
+            'smb_lookupsid=patator.patator:main',
+            'rlogin_login=patator.patator:main',
+            'vmauthd_login=patator.patator:main',
+            'mssql_login=patator.patator:main',
+            'oracle_login=patator.patator:main',
+            'mysql_login=patator.patator:main',
+            'mysql_query=patator.patator:main',
+            'rdp_login=patator.patator:main',
+            'pgsql_login=patator.patator:main',
+            'vnc_login=patator.patator:main',
+            'dns_forward=patator.patator:main',
+            'dns_reverse=patator.patator:main',
+            'snmp_login=patator.patator:main',
+            'ike_enum=patator.patator:main',
+            'unzip_pass=patator.patator:main',
+            'keystore_pass=patator.patator:main',
+            'sqlcipher_pass=patator.patator:main',
+            'umbraco_crack=patator.patator:main',
+            'tcp_fuzz=patator.patator:main',
+            'dummy_test=patator.patator:main',
         ],
     },
 )
